Rename teleop package to flight_control
