+++ src/surface/flight_control/doc/images/SelectorNodeGraph.png
