+++ src/surface/flight_control/flight_control/auto_docking_node.py
