+++ src/surface/flight_control/flight_control/manual_control_node.py
@@ -11,7 +11,7 @@ from mavros_msgs.msg import OverrideRCIn
 
 from rov_msgs.msg import CameraControllerSwitch, Manip
 
-from teleop.pixhawk_instruction import PixhawkInstruction
+from flight_control.pixhawk_instruction import PixhawkInstruction
 
 # Button meanings for PS5 Control might be different for others
 X_BUTTON:        int = 0  # Manipulator 0
