+++ src/surface/flight_control/launch/flight_control_launch.py
@@ -4,7 +4,7 @@ from launch_ros.actions import Node
 
 def generate_launch_description() -> LaunchDescription:
     manual_control_node: Node = Node(
-        package='teleop',
+        package='flight_control',
         executable='manual_control_node',
         remappings=[('/surface/manipulator_control', '/tether/manipulator_control'),
                     ('/surface/mavros/rc/override', '/tether/mavros/rc/override')],
@@ -13,7 +13,7 @@ def generate_launch_description() -> LaunchDescription:
     )
 
     auto_docking_node: Node = Node(
-        package='teleop',
+        package='flight_control',
         executable='auto_docking_node',
         remappings=[('/surface/manipulator_control', '/tether/manipulator_control'),
                     ('/surface/mavros/rc/override', '/tether/mavros/rc/override')],
