+++ src/surface/flight_control/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>teleop</name>
+  <name>flight_control</name>
   <version>1.0.0</version>
   <description>Mate ROV sub movement controllers</description>
   <maintainer email="bwp18@case.edu">Benjamin</maintainer>
