+++ src/surface/flight_control/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/flight_control
+[install]
+install_scripts=$base/lib/flight_control
