--- src/surface/teleop/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/teleop
-[install]
-install_scripts=$base/lib/teleop
